fixed problems with grid layout
