+++ src/assets/cgmpresser.js
@@ -1,0 +1,7 @@
+export const cgmpresser = {
+    "name": "Clif",
+    "gif": "https://media.giphy.com/media/3oEjHLzm4BCF8zfPy0/giphy.gif",
+    "skills": "Java, JavaScript, C++, SQL (skills I've forgotten would be a much longer list)",
+    "github": "https://github.com/cgmpresser"
+}
+export default cgmpresser
