+++ src/assets/index.js
@@ -1,1 +1,2 @@
 export * from './thedjziegler'
+export * from './cgmpresser'
