+++ src/assets/thedjziegler.js
@@ -1,5 +1,5 @@
 export const thedjziegler = {
-    "name": "Danny Ziegler",
+    "name": "Danny Z",
     "gif": "https://media.giphy.com/media/lp3GUtG2waC88/giphy.gif", // using the gif link itself
     "skills": "Java, Javascript, Python",
     "github": "https://github.com/TheDJZiegler"
